bot/pyproject.toml
[project]
name = "moirai-bot"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "aiogram>=3.4,<4",
    "aiohttp_socks>=0.10,<1",
    "pydantic-settings>=2,<3",
    "google-api-python-client>=2.130,<3",
    "google-auth>=2.30,<3",
]

[project.optional-dependencies]
dev = [
    "ruff",
    "pytest>=8,<9",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["moirai_bot"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]

[tool.ruff.format]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
